chromeos-config: cache the results of GetValidSchemaProperties
This function is called once-per-config during the creation of a
CrosConfigJson object.  On programs like octopus, there is greater
than 130 configs for the single program, and this function is costly
as we have to re-parse yaml every single execution.

Caching the results of this function speeds the execution of
get-firmware-build-combinations on octopus from 27.3 seconds to 3.5
seconds.

BUG=chromium:1111055
TEST=speedup on octopus mentioned above

Change-Id: Id3cae6eeae8edb46b3456f72fdb05d4e9768ba69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2335264
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>