chromeos-config: move the retry logic outwards for creating loop devices
In CL:2320013, the initialization of cros_config was moved earlier
into the boot, blocking boot-services so that services like the
fingerprint firmware updater can use cros_config.

However, this exposed another race condition in cros_config: during
early boot, even if /dev/loop-control can be opened and we get a loop
file number, the device may still be busy, either by use from another
process or by the kernel.

Unfortunately, in this situation, just like with /dev/loop-control,
the only option is to retry the operation until we are able to
complete it: there is no option to block on this file until it's
ready.

We percolate the existing retry mechanism outwards to surround the
entire loop setup, and check for busy errors to do the retry.

NOTE: this change needs cherry-picked to M85... I appreciate any
prioritization in the review you are able to do.

BUG=b:162749841,b:162720284,b:162541620,b:162719399
TEST=on eve, 30/30 reboots has working cros_config (compared to
     failure with only 2-3 reboots before this change)

Change-Id: I1eada39c2ed8614a1cb4d4c85bc06cd9fe3fa345
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2335706
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>