ocr: Add OCR service Mojo adapter
Add ocr_service_mojo_adapter that provides a Mojo connection
to the OCR daemon, which allows other processes to call the
Mojo interface methods. It will be used in the CLI which will
be developed for running integration tests.

The adapter uses a delegate to bootstrap the Mojo connection
via D-Bus.

BUG=b:159229207
TEST=None

Change-Id: Ia72851a129ae2f3090475bff07c5c5abff92f977
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2337054
Tested-by: Efi Mavroudi <emavroudi@google.com>
Commit-Queue: Jesse Schettler <jschettler@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>