shill: persist WiFi.GlobalFTEnabled to the default profile
This Manager property is used to enable or disable FT and can be
configured by enterprise policy. It should be persisted to the default
profile so that the property is not reset to the default value on shill
restart.

We change ft_enabled to be a base::Optional<bool> because we need a way
to differentiate between the default value as defined by the DISABLE_FT
flag and the value set by the enterprise policy via the Manager
property. The property is only persisted if it is set by D-Bus,
otherwise flipping the DISABLE_FT switch to change the default value
won't do anything since the value will always be loaded from the default
profile.

BUG=b:158665656
TEST=Use dbus-send to set the property, restart shill, check that
/var/cache/shill/default.profile contains the expected value, and that
GetProperties also reflects the expected value. Clear the property and
make sure that it reverts to the default value (true when
USE=fast_transition is enabled, false otherwise). Also unittests.

Change-Id: I3876886d1897692bc90114bf1b12e83079b43fba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2337314
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>