diagnostics: Fix system info fuzzer
Edit the fuzzer to write given input data to the files that the system
info will read from. Add constraints for extremely large input size.
Previously, the fuzzer, running on the workstation, took ~28000 ms to
complete on the failing test case (which can be found in the linked
BUG), likely causing a timeout on slower machines that are not as not as
powerful as workstations. With this change, the fuzzer is running
successfully in less than ~30 ms.

BUG=chromium:1112106
TEST="/usr/libexec/fuzzers/fetch_system_info_fuzzer
/tmp/clusterfuzz-testcase-minimized-fetch_system_info_fuzzer-5846721078296576"
"/usr/libexec/fuzzers/fetch_system_info_fuzzer
/tmp/clusterfuzz-testcase-minimized-fetch_system_info_fuzzer-6029567189385216"

Change-Id: I7a1dbdf647c0af23fa2c3e45abc8d08c4390a480
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2337818
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Kartik Hegde <khegde@chromium.org>
Tested-by: Kartik Hegde <khegde@chromium.org>