tpm_manager: avoid unnecessary default auth test
Previously the default auth test is performed w/o any aid of the local
data in persistent storage. Thus it causes DA to increase every time
tpm_manager starts. For those device that is incapable of resetting DA,
the consequence is that DA gets locked out and users cannot log in.

If the device has the flag to tell accomplished tpm initialization, the
default auth test would just skip so the DA counter doesn't increase. It
should help withe cases of missing delegate or delegate w/o DA reset
permission.

Note that in case of DA lockout in effect, this CL does very little
because checking the TPM ownership itself also fails in that situation.

To test the CL, on a real device a delegate is created w/o permission of
resetting DA. This setup is verified by observing failed DA reset after
clearing owner password. Then, in sequence --
1. Clear the ownership, no DA increment, just to be sure.
2. Manually touch the flag file of interest, restart tpm manager, and
observe the flag is wiped.
3. Take ownership, and obseve the flag shows up.
4. Manurally remove the flag file of interest, restart tpm manger, and
observe that DA increment by 1 and the flag file of interest shows up.
5. restart tpm manager, no further DA increment.

In case of default password, it is verified tpm maanger doesn't set the
flag file, and of course it doesn't increase the DA counter.

BUG=chromium:1110741
TEST=See above.

Change-Id: I7f6fd0855442db7b47c8726c647c52fab661de08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2340830
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Leo Lai <cylai@google.com>