vm_tools: sommelier: Set dmabuf modifier as invalid.
Sommelier was setting the modifier for dmabufs to be 0 when it was not
actually known.  This was causing exo/Chrome to interpret buffers from
Crostini incorrectly (tiled vs linear).  This is not a full/complete
fix -- the modifier should be queried from the BO and passed back to
Chrome but currently there isn't support plumbed through for that.

BUG=b:162797079
TEST=glxgears

Change-Id: I343c25963e42d23eaecc988a2b97e8f08f0b1a02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2339286
Tested-by: David Riley <davidriley@chromium.org>
Auto-Submit: David Riley <davidriley@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>