arc: vm: libvda: fix invalid consistency check in InitDecodeSession()
connection_ is set by the GpuVdaImpl constructor, so it will always be
valid no matter when we call this method. ipc_task_runner_ is the member
that is set by Initialize() and thus the one we should check to make
sure we have been called after it.

Furthermore, since the threading DCHECK() also relies on
ipc_task_runner_'s availability, move it after that check.

BUG=None
TEST=arc.LibvdaDecode.h264 passes on Kukui.

Change-Id: I84ea461724e19ef3cac5dd26f660cd0cf1c4247c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2340667
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Alex Lau <alexlau@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>