arc: vm: libvda: fix potential crash during test cleanup
When running all the tests of libvda_gpu_unittest, a crash sometimes
occurred:

    Check failed: ptr_.
    Task trace:
    /usr/local/libexec/libvda-gpu-tests/libvda_gpu_unittest(+0x179e9) [0x10809e9]
    Aborted (core dumped)

This seems to be due to the fact that some destruction tasks are posted
on the IPC thread before a decoding session is destroyed, and when
finally run the task operates on invalid data.

Fix this by reusing the RunTaskOnThread() function from
vaf_connection.cc to wait until the destroy method completes, in order
to make sure that all state is cleaned before we return.

Interestingly this issue was not spotted by arc.LibvdaGpuTests since it
runs the 4 tests individually.

BUG=b:160754649
TEST="/usr/local/libexec/libvda-gpu-tests/libvda_gpu_unittest --test_video_file=/usr/local/share/tast/data_pushed/chromiumos/tast/local/bundles/cros/arc/data/test-25fps.h264 --gtest_repeat=1000" passes on Kukui.

Change-Id: Iad8e4ef5977baf5c0cafa106d0f427ab709d4edf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2341271
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Alex Lau <alexlau@chromium.org>