typecd: Add cable PD ID check for TBT3 alt mode
According to the Type-C Specification (Rel 2.0) Figure F-1, an important
part of the TBT3 alt mode entry flow is to check the PD Identity Header
of the Cable for appropriate speed support.

Add that check, and add a unit test to verify it works as expected.

BUG=b:152251292
TEST=P2_TEST_FILTER="*.*" cros_workon_make --board=volteer --test
typecd passes.

Change-Id: I593bde3e9960972ac884e1fdb941261399f38f00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2348252
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>