debugd: Support possible binary /var/log/messages in grep
Sometimes the /var/log/messages file gets corrupted and some non-ascii
text is there. That makes this grep fail to print anything out unless we
pass -a to it. Let's do that so we can find touch firmware versions in
messages files that are a little binary.

BUG=None
TEST=Add some binary to messages, reboot, check chrome://system to see
touch firmware version

Change-Id: Ia0d47f0f80495b05f276621b0080cbb3049eac11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2349648
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>