system_api: Fix integer type in VmInfo
vm_tools.concierge.VmInfo.seneschal_server_handle is defined as
uint64, but seneschal itself uses 32 bit handles. This seems to be
handled correctly inside of concierge itself, it's just the protobuf
definition that's wrong.

According to the language guide uint64 and uint32 are compatible types
so this won't cause parsing errors, and nothing larger then the
maximum value of a uint32 will be stored in this field in any case.

BUG=none
TEST=CQ

Change-Id: I6945a777de0259e24e1e71302acc1e0d96dd10e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2351200
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>