vm_tools: vsh: Register vsh and container shell pid with cicerone
When vsh starts a new session, vshd returns the pid of the forked shell
in the container and vsh registers the host vsh pid and container shell
pid with cicierone.

vsh also now accepts --cwd to be a pid and passes this value to vshd.
vsh_forwarder will lookup /proc/<pid>/cwd and start the new shell in
the same cwd.

When terminal launches a new session, it can query cicerone for the
container shell pid, and then use it when launching the new vsh to have
the new session use the same cwd.

BUG=chromium:1113207
TEST=vsh ...--cwd=<pid>

Change-Id: I7f051c1aeed81d70fd20dab4efb1f7da01da5633
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2351207
Tested-by: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>