chromeos-config: re-add support for customization-id on ARM
We dropped support for the customiztion ID on ARM as no unibuild ARM
devices were old enough to have the customization ID.

But now that we are migrating old devices, we need this again,
specifically to identify Hana zergs.

Add support back to the schema and make necessary code changes to
support this.

BUG=chromium:1115620
TEST=unit tests

Change-Id: Ic293d08a45df421ee3f7efed7468ac8179eeaedf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2352435
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>