power: Listen to udev events to detect peripheral battery change
Currently powerd polls peripheral battery information every 10 minutes.
This has been fine since we only use it for low-battery notification.
Now that we use this to display the battery value to the UI, the update
signal needs to be more timely when a change is detected about a
peripheral device.

This patch makes PeripheralBatteryWatcher also listen to udev events so
that it can update a peripheral battery status immediately when a change
is detected, before the next periodic poll happens.

BUG=b:164167716
TEST=Added unit tests and ran: FEATURES=test emerge-$BOARD power_manager
Tested manually on Chromebook with Apple Magic Mouse:
* Disconnect the mouse.
* Start dbus-monitor.
* Connect the mouse.
* Observe in the dbus-monitor that a signal is immediately emitted by
  powerd about the newly connected mouse's battery.

Change-Id: I43ead4fe9123386fa5ac79437279eca9d2d38a1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2354984
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Sonny Sasaka <sonnysasaka@chromium.org>