crash: arc_collector: Implement a core_collector64 and use it
If we're running a 32-bit chrome userspace and a 64-bit arc++ container
then we need to collect core dumps for both 32-bit and 64-bit processes,
similar to x86 builds. Implement the logic to collect the core dump from
arc++ properly when it contains a mix of 32-bit and 64-bit code.

BUG=b:164481201
TEST=kill -6 <pid of hwcomposer in container>, see minidump created in
     /home/chronos/crash/

Cq-Depend: chromium:2358621
Change-Id: Ifc36623a942e905d05dc6b466b7cff6e5940fa32
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2358622
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>