lorgnette: read existing SANE option values
In SaneClientImpl's LoadOptions function, read the existing option
values as they are found, so that they can be returned to the client
if requested.

Convert SaneOption to use a std::vector<char> as the backing memory for
string options. This replaces uses of calloc/free, which are extremely
error prone.

BUG=b:161239919
TEST=lorgnette_cli --scan_source="ADF Simplex"

Change-Id: I7ad67041111ab900c9909441af6bfc313b0cf444
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2361030
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>