shill: Allow Cellular auto connect on Cellular enable
This clears the "explicit disconnect" bit when Cellular
services are loaded, which occurs when a SIM changes or when
a device is enabled.

This will allow Cellular services to auto connect after a
disable/enabler or after an eSIM profile change, even if they
were explicitly disabled previously.

BUG=b:163379070
TEST=manual (see issue)

Change-Id: Ib1ff3c8d41754570113017ac5ebd92ad9bc3540d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2360927
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>