camera: hal: ip: Fix race-condition when disconnecting a running camera
The camera HAL adapter assumes that the device won't make any more
callbacks after camera_device_status_change() is called indicating that
the device is disconnected. Calling Close() on the device generates some
callbacks when it flushes the queue. In order to avoid this race, call
Close() on the device before making the camera_device_status_change()
call.

Also add a check to ensure that a capture request sent to a device that
has had Close() called on it will fail.

BUG=b:165034331
TEST=Manually tested closing a running camera with sleep() calls added
in the code to trigger the race condition.

Change-Id: I15b53bf39dc4e628a5170447b29ad51affa0196c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2360607
Tested-by: Pablo Ceballos <pceballos@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>