crash: arc_collector: Clean up core_collector32 support
We shouldn't need this #if here because we always have a
core_collector32 when this program is compiled for 64-bit and some
32-bit process is crashing. Clean up the comment in the Makefile about
PIE linking as it has bit rotten now that the eclass is the one doing
the static linking. And finally, remove the TODO because the bug is
closed now and we're not going to make core_collector be
core_collector32 on arm32 builds.

BUG=b:164481201
TEST=CQ passes

Change-Id: I2d0aca33a2997e3c30467205f615254af5d07a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2361663
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>