diagnostics: add RunMemtester() method to executor
This new method runs the memtester executable for one loop on all
of the device's free memory, except for 200MB. This behavior is
consistent with crosh's memory_test command. The executor needs to
run memtester because it requires capabilites to mlock memory.

BUG=chromium:1113897
TEST=emerge-nami diagnostics

Cq-Depend: chromium:2352621
Change-Id: I7f25ad8a424c435f3ac332dd73e17a0c2298f157
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2363428
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>