modemfwd: clear LTE attach APN
Workaround proposed by Fibocom.

When the user is swapping the SIM on a running system, the modem might
try to connect to the new carrier network with the old attach APN.
Some carrier just fails the connection in the case.

Let the modem specific helper the possibility of clearing its
non-volatile memory for this specific case.

BUG=b:163476637
TEST=run on Jinlon and check the modemfwd log when swapping SIMs.
TEST=FEATURES=test emerge-hatch modemfwd

Change-Id: Iff4db88cf2e4d3146ff3edd0b9f9e3d704030ba6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2368356
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Henry Sun <henrysun@google.com>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>