shill: Restore Service.DHCPProperty
go/jetstream relies on Service.DHCPProperty.VendorClass to set
different values for different active Ethernet services.

That functionality was reverted here, as it is unused in Chrome OS:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2347138

This CL restores that functionality and adds comments to
service-api.txt and service.h to ensure this is not removed
without checking with the jetstream project going forward.

BUG=b:165648891
TEST=DhcpPropertiesTest.*, ServiceTest.(Load|Save)

Change-Id: I58aa7f04dc9a80db41affb6bbd4aa57d0b0934db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2367563
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>