crash: kernel_collector: Fix signatures for x86_64 kernels
Newer x86 kernels aren't printing the RIP address. Instead, they just
print the symbol of the function and the offset inside that function.
This was printed before in other bug reports, but we used a later RIP
print that had a similar signature (see the
ComputeKernelStackSignatureX86_64 test). Update the regex for x86_64 so
that it matches newer and older kernels and add in a test for the new
style.

BUG=chromium:1117375, b:165884963
TEST=tast run $DUT crash.KernelCrash.*

Change-Id: Iaea640b65fba281077783f7ac277f8caf26c3733
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2368310
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>