patchpanel: define USE flag for disabling some routing features
The Jetstream project based on Chrome OS is not running hosted VMs or
containers and needs a way to disable the associated routing features
and to disable patchpanel adding fwmark and forwarding rules in
iptables/ip6tables.

This patch introduces a build flag USE_JETSTREAM_ROUTING to do so. It
is off by default for Chrome OS.

BUG=b:162966185
TEST=Compiled patchpanel with USE=jestream_routing, deployed on
rammus, rebooted, checked that fwmark rules are not set up and ARC
connectivity does not work anymore.

Change-Id: Ia8152a58c2f87a627fef9cfa1a521bbf0419f32e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2369078
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>