camera: libcamera_connector: Revamp synchronization
Previously we made sure all StartCapture(), StopCapture() and
SendCaptureResult() calls are atomic with a centralized lock. In this
CL, we completely revamp the original method with a sequenced approach.

New synchronization mechanism:
- Make the previously-mentioned calls run on the same sequenced task
runner.
- Define a task, |SessionRequest|, which stores
  - Whether the request is to start or stop a capture session
  - Information about the session, e.g., camera ID, format, etc.
  - Future callback to execute once the task is finished.
- Declare a task queue and push StartCapture() and StopCapture() in the
queue if we're busy with ongoing session requests.
- Consume tasks when we finish opening or closing a device.

This approach allows for more finegrained control on the state
transitions (e.g., what happens when user stops a session that's already
stopping). To do this, we create a state machine with the following
states: kIdle, kStarting, kCapturing, kStopping, then we assess probable
state transitions and define the treatments for each transition in
StartCaptureOnThread() and StopCaptureOnThread().

BUG=b:151047930
TEST=cros_camera_connector_test && cros_camera_connector_client

Change-Id: I9cd186da47aa658086fc36a210d44b78c4e172c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2369091
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Jasmine Chen <lnishan@google.com>
Tested-by: Jasmine Chen <lnishan@google.com>