authpolicy: Fix for setting fake domain SID.
In the latest Samba versions (4.8.6+), "net gpo list" expects
domain SID to be set, which is s associated with a workgroup
in secrets.tdb.

The previous code was setting domain SID only once in authpolicy
daemon lifetime associating it with the first workgroup for which
MaybeSetFakeDomainSid() was called. Therefore, it was failing if
workgroup was different for user and device accounts.
This fix is making sure setdomainsid is called for every account
workgroup.

BUG=chromium:1121606
TEST=emerge-${BOARD} samba && test_that -b ${BOARD} ${IP_OF_DUT} enterprise_AuthPolicyDaemonServer.auth

Change-Id: I81341c5bf8804146f42ac6c138c23a1fc00304ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2375331
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Tested-by: Tomasz Dobrowolski <tomdobro@chromium.org>
Commit-Queue: Tomasz Dobrowolski <tomdobro@chromium.org>
Auto-Submit: Tomasz Dobrowolski <tomdobro@chromium.org>