permission_broker: deny detach of driver from fixed USB devices
Some boards have internal USB-attached devices (webcam, Bluetooth
controller, etc.), which should not be treated as normal USB devices
from the point of view of the user.  In particular, it should not be
possible to detach them from their drivers, as this would cause loss of
functionality for built-in devices.

Use the "removable" sysfs attribute to determine whether a device is
built into the system.  Not all systems provide this information, so
this check only excludes devices that are explicitly reported as
"fixed"; devices that are "removable" or "unknown" follow the existing
detach/lockdown logic.

BUG=None
TEST=Verify 'Denying fixed USB device' messages for camera on kohaku

Change-Id: I318968f77a8e94a365a87fc12c897c5f2fd13b3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2378554
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>