biod: Breakup cros_fp_updater
The cros_fp_updater{,_test}.cc files are massive and
confusing. Some of this functionality needs to be moved
out of these files in order to be more manageable and
independently testable.

Some of the functionality relates to cros_config, which
can be used by other biod elements. For this reason,
we add biod_config, which is intended to hold cros_config
and other configuration mechanisms/settings.

Some of this functionality relates specifically to the
updater, but can be separated from the main updater logic.
These functions have been relocated to updater/update_utils*.

All logic remains the same after this patch, except for
the FindFirmwareFile function, where it was converted to use
biod::FingerprintBoard with an Optional return value.

BUG=b:153881504
TEST=emerge-hatch biod
TEST=FEATURES=test emerge-hatch biod

Change-Id: I3ba801f2ce73c403ddc5c5dc258f86e97ffd33cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2376827
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>