power: Add API to refresh a Bluetooth device battery signal
The battery information was designed to be a periodic signal. This was
fine because the need was for low battery warning. Now that we want to
display the Bluetooth battery values continuously, there is a problem
where Chrome does not have the battery values after Chrome restarts (for
example when signing out) until powerd's next battery poll.

A workaround for this is to have an API in powerd for Chrome to request
a signal for a Bluetooth device. This lets Chrome sync with powerd's
battery values when Chrome starts up. This workaround will be removed
when we migrate to go/bluez-battery-provider-api.

BUG=b:166530001
TEST=Unit test:
* FEATURES=test emerge-$BOARD power_manager
Manual test:
* Apply the corresponding Chrome change CL:2378788
* Connect a Logitech M535 to DUT
* Notice that the battery value is displayed in system tray
* Sign out
* Check that the battery value is available right away in system tray

Cq-Depend: chromium:2379152
Change-Id: Id66b6294697d40263f8d648d5c9f6d77b72e421b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2378569
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Alain Michaud <alainm@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Alain Michaud <alainm@chromium.org>