cros-disks: [partition-as-single] Add PartitionManager implementation.
It provides PartitionAsSingle service to wipe exiting partitions and
create a new single partition on removable disk.

It uses sfdisk command to partition the disk. It's called through a
sandbox to restrict access to resources.

The default partition type is mbr (MS_DOS) unless the disk
size is >2TB that will be gpt.

BUG=chromium:491043
TEST=unit test

Change-Id: If6a77c6d388d60cf0429c94bf16ab7d4c4688e57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2379389
Tested-by: Omid Tourzan <oto@chromium.org>
Commit-Queue: Omid Tourzan <oto@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>