biod: Allow biod_storage_fuzzer to take non-UTF8 bytes as data
The data (template) bytes are not necessarily UTF8, so we should allow
the fuzzer to take non-UTF8 bytes in the data field. Previously this
field was passed to base::Value's constructor as a string, and that
resulted in a IsStringUTF8 check that we don't want.

Pass the data bytes as std::vector<uint8_t> so that base::Value
constructor does not complain.

BUG=b:162871040
TEST=cros_fuzz --board=hatch reproduce --fuzzer biod_storage_fuzzer
     --testcase /build/hatch/tmp/clusterfuzz-testcase-minimized-biod_storage_fuzzer-5179296473219072
     --package biod --build-type asan

Change-Id: Id5c04fe2d192c695ab52fc4c7afc6bd84e5080c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2382396
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Auto-Submit: Yicheng Li <yichengli@chromium.org>