lorgnette: delete ScanJobState when scan fails
If a multi-page scan fails, delete the ScanJobState (which holds a
std::unique_ptr<SaneDevice>) for the scan.

This fixes an issue where a device is unusable after a failed scan until
Lorgnette automatically closes.

Also refactor the release_device ScopedClosureRunner so that if the
device has been deleted, it does nothing. This makes the signature and
logic of GetNextImageInternal a little simpler.

BUG=None
TEST=run ADF scan with empty document feeder, then run scan again.

Change-Id: Idbf8e1767aa44b64b8faf25c7ea008c3ec47e295
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2386497
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>