vm_tools: concierge: add a name encoding utillity.
It turns out that vm name encoding is nonstandard, some places it is
padded and some it is not.

We fix this by moving the encoding to a utility function. We choose
padding over non-padding as currently all the unpadded names can be
re-generated as needed but the padded one is the VM stateful disk which
is harder to move.

We make this utility common (which prompts a minor refactoring of the
build) as it is used both by concierge and vm_syslog.

BUG=b:167294302
TEST=run crostini, view container contents in the files app.

Cq-Depend: chromium:2392220
Change-Id: If2e5365111a02f9fe19279f7de9b617016b71c2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2386885
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>