shill: Support rollbacks for simple WiFi networks
This CL preserves the rot47 Passphrase property with
comments to maintain it through M89, after which rollback
support will end for M85.

This also adds comments to ensure that rollback supported
properties are not changed in a rollback incompatible way.
See issue for details.

BUG=chromium:1120161
TEST=manual (see issue)

Change-Id: I2f471f5e6ac0c7981da253a0dd71ae0155f0eb0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2382304
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>