cryptohome: Finalize bootlockbox before creating Mount
Mount process doesn't need to know about BootLockbox, so ensure it is
finalized before we create Mount object and factor it out from Mount.
In future, if no better place for that is found, the MountFactory shall
be responsible for doing that. Not doing it now as the hierarchy of the
classes may change a lot soon.

BUG=chromium:1124561
TEST=unittest

Change-Id: Ide35dca592ff5897bd4b38a4c7cf69781c74e8c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2390605
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Xiaoyong Zhou <xzhou@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>