chromeos-common-script: Use largest nvme namespace for get_fixed_dst_drive
NVMe devices could potentially have more than one namespace. We
need to make sure that we are using the largest namespace for
chromeos-install. This means that "nvme0n1" in rootdev_base will
still be present, as that is the expected rootdev if there is only
one namespace on the device, but it will be overloaded if there is
a bigger namespace on the SSD.

BUG=chromium:1124582
TEST=chromeos-install on 2 namespace device, detected correct namespace
as dst

Change-Id: Iff394e9b7f9d78a51906e09051c3a74a86c7c3ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2392143
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>