patchpanel: Change SocketForwarder exit condition
This CL replaces the exit condition of the SocketForwarder from quitting
as soon as one of the sockets receives a EPOLLRDHUP signal to waiting
for both sockets to receive EOF.

EPOLLRDHUP means that the peer will no longer write to the socket, but
they can still receive data on that socket. In practice, this is
encountered when downloading PlayStore apps through a SocketForwarder.

BUG=chromium:1125177
TEST=downloading PlayStore apps on DUT works

Change-Id: I387cdf413e4a0ececd4453d0783424388de145d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2398698
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>