lorgnette: Map ippusb backend to airscan
A "backend" of ippusb indicates that we want to connect to an eSCL
scanner through an IPP-over-USB tunnel.  When this backend is requested,
contact ippusb_manager to find the device and create the tunnel, then
rewrite the device string to pass the local socket to the airscan
backend.

BUG=b:153544296
TEST=Scan from USB-connected scanners

Change-Id: Ie0b07d191c821d7ab2489f04f4b8f20dcde39690
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2404084
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>