crash: Collect SMMU faults in the kernel warning collector
Sometimes there are SMMU faults where a device is accessing an address
that isn't mapped for it. Let's catch these sorts of problems and send
them to crash so they can be analyzed. Add support for this kernel
message to crash and some tests to make sure the implementation is sane.

BUG=None
TEST=FEATURES="test" emerge-${BOARD} crash-reporter

Change-Id: I48edb57e83cee90a90af27e373adf2714f9c08bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2399529
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>