vm_tools: vmc can calculate hashes in multi-user scenarios
If CROS_USER_ID_HASH isn't set vmc will try to figure it out. Previously
it only worked when there was only a single account on the device, but
we can figure out the hash of the currently logged-in user even when
there are multiple accounts on the machine.

BUG=None
TEST=vmc start termina

Change-Id: I6aed745ac3d88c65f0e356bd3ad28cd87de9df94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2398354
Tested-by: David Munro <davidmunro@google.com>
Auto-Submit: David Munro <davidmunro@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: David Munro <davidmunro@google.com>