debugd: do not retain SELinux label on log
When copying the log from before encrypted partition is mounted,
do not try to retain the SELinux label. It causes the "tmpfs" label
to want to be written to the filesystem which is invalid for
labeled filesystems.

See crrev.com/c/887217 for the same issue with cryptohome.

BUG=chromium:887859
TEST=ls -lZ /var/log/clobber-state.log

Change-Id: I467e4145ce10dbb479b3d7aeb310eb235d06505d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2399755
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>