debugd: accept username in arc backup apis
Chrome code shouldn't be sending the hash to debugd directly since it is
a cros concept and shouldn't know how to hash the account id (which it
is doing at the moment). So this change would support sending account id
from chrome.

In this change we update the BackupArcBugReport and
DeleteArcBugReportBackup method of debugd to take both the username and
userhash and in the implementation we check if it is a valid hash, use
it as it is otherwise hash it and then use it.

We will then update all the clients to send username instead of userhash
and then remove the support for userhash being sent directly.

BUG=b:168153061
TEST=cros_run_unit_tests --board eve --packages debugd
TEST=cros_run_unit_tests --board eve --packages chromeos-login
TEST=dbus-send --system --dest=org.chromium.SessionManager
               --type=method_call --print-reply --fixed /org/chromium/SessionManager
               org.chromium.SessionManagerInterface.StopArcInstance
               string:'<username>' boolean:true
TEST=dbus-send --system --dest=org.chromium.debugd \
                 --type=method_call \
                 --print-reply --fixed \
                 /org/chromium/debugd \
                 org.chromium.debugd.BackupArcBugReport \
                 string:'<username>'
TEST=dbus-send --system --dest=org.chromium.debugd \
                 --type=method_call \
                 --print-reply --fixed \
                 /org/chromium/debugd \
                 org.chromium.debugd.DeleteArcBugReportBackup \
                 string:'<userhash>'

Change-Id: I7129a377aa5c010efb4d7bdf2851e3d892bbbd6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2402246
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>